Fixed image width issue + removed mso line height rule
Images were not stretching to the full width of their parent element. I updated the internal styling on the image a tags, image tags, and changed the margin, padding and border reset to affect all elements rather than just the table and div elements. This together solved the image spacing issue.